pyproject.toml
[project]
name = "x-tweet-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "ruff>=0.14.13",
    "uiautodev>=0.14.0",
    "uiautomator2>=3.5.0",
]
